cri: retry stop container if there is connection closed
In critest, a lot of containers will exit automatically. It could be
race condition between event handler and StopPodSandbox goroutines. The
StopPodSandbox could encounter `ttrpc: closed` error. It's retryable. To
fix flaky CI issue, add new function as wrapper to retry stopContainer
call.
