+++ internal/cri/server/podsandbox/sandbox_stop.go
@@ -47,14 +47,7 @@ func (c *Controller) Stop(ctx context.Context, sandboxID string, _ ...sandbox.St
 	}
 	state := podSandbox.Status.Get().State
 	if state == sandboxstore.StateReady || state == sandboxstore.StateUnknown {
-		err = c.stopSandboxContainer(ctx, podSandbox)
-		if err != nil && isShimTTRPCClosed(err) {
-			log.G(ctx).WithError(err).
-				Warnf("Shim ttrpc connection closed when stopping sandbox container %q in %q state, retrying", sandboxID, state)
-
-			err = c.stopSandboxContainer(ctx, podSandbox)
-		}
-		if err != nil {
+		if err := c.stopSandboxContainerRetryOnConnectionClosed(ctx, podSandbox); err != nil {
 			return fmt.Errorf("failed to stop sandbox container %q in %q state: %w", sandboxID, state, err)
 		}
 	}
@@ -62,6 +55,30 @@ func (c *Controller) Stop(ctx context.Context, sandboxID string, _ ...sandbox.St
 		return fmt.Errorf("failed to cleanup sandbox files: %w", err)
 	}
 	return nil
+}
+
+func (c *Controller) stopSandboxContainerRetryOnConnectionClosed(ctx context.Context, podSandbox *types.PodSandbox) error {
+	const maxRetries = 3
+
+	var err error
+	for i := 1; i <= maxRetries; i++ {
+		err = c.stopSandboxContainer(ctx, podSandbox)
+		if err == nil {
+			return nil
+		}
+
+		if !ctrdutil.IsShimTTRPCClosed(err) {
+			return err
+		}
+
+		if i+1 <= maxRetries {
+			retryAfter := time.Duration(100*i*i) * time.Millisecond
+			log.G(ctx).WithError(err).
+				Warnf("Shim ttrpc connection closed when stopping sandbox container %q, retry after %s", podSandbox.ID, retryAfter)
+			time.Sleep(retryAfter)
+		}
+	}
+	return err
 }
 
 // stopSandboxContainer kills the sandbox container.
